Refactor(FileDropzone): restructure component and enhance functionality
- Moved FileDropzone to a new directory and split it into smaller components: Loader, Preview, and StartUpload for better organization and maintainability.
- Updated translation resources to include a new loading message.
- Adjusted related components and routes to accommodate changes in props and state management.
